#703 fix the error to insttanse obj
#703 fix the error to insttanse obj
